Move writeFactory to util.dart
