+++ json_serializable/lib/src/json_serializable_generator.dart
@@ -8,7 +8,6 @@ import 'dart:collection';
 import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/type.dart';
 
-import 'package:analyzer/analyzer.dart';
 import 'package:json_annotation/json_annotation.dart';
 import 'package:source_gen/source_gen.dart';
 
@@ -122,7 +121,7 @@ class JsonSerializableGenerator
           _deserializeForField(field, classAnnotation.nullable,
               ctorParam: ctorParam);
 
-      var fieldsSetByFactory = _writeFactory(
+      var fieldsSetByFactory = writeFactory(
           buffer, classElement, fields, ignoreReason, deserializeFun);
 
       // If there are fields that are final – that are not set via the generated
@@ -458,115 +457,4 @@ InvalidGenerationSourceError _createInvalidGenerationError(
 
   return new InvalidGenerationSourceError(message,
       todo: 'Make sure all of the types are serializable.');
-}
-
-/// Returns the set of fields that are written to via factory.
-/// Includes final fields that are written to in the constructor parameter list
-/// Excludes remaining final fields, as they can't be set in the factory body
-/// and shouldn't generated with toJson
-Set<FieldElement> _writeFactory(
-    StringBuffer buffer,
-    ClassElement classElement,
-    Map<String, FieldElement> fields,
-    String ignoreReason(String fieldName),
-    String deserializeForField(FieldElement field,
-        {ParameterElement ctorParam})) {
-  var fieldsSetByFactory = new Set<FieldElement>();
-  var className = classElement.displayName;
-  // Create the factory method
-
-  // Get the default constructor
-  // TODO: allow overriding the ctor used for the factory
-  var ctor = classElement.unnamedConstructor;
-  if (ctor == null) {
-    throw new UnsupportedError(
-        'The class `${classElement.name}` has no default constructor.');
-  }
-
-  var ctorArguments = <ParameterElement>[];
-  var ctorNamedArguments = <ParameterElement>[];
-
-  for (var arg in ctor.parameters) {
-    var field = fields[arg.name];
-
-    if (field == null) {
-      // ignore: deprecated_member_use
-      if (arg.parameterKind == ParameterKind.REQUIRED) {
-        var msg = 'Cannot populate the required constructor '
-            'argument: ${arg.displayName}.';
-
-        var additionalInfo = ignoreReason(arg.name);
-
-        if (additionalInfo != null) {
-          msg = '$msg $additionalInfo';
-        }
-
-        throw new UnsupportedError(msg);
-      }
-
-      continue;
-    }
-
-    // TODO: validate that the types match!
-    // ignore: deprecated_member_use
-    if (arg.parameterKind == ParameterKind.NAMED) {
-      ctorNamedArguments.add(arg);
-    } else {
-      ctorArguments.add(arg);
-    }
-    fieldsSetByFactory.add(field);
-  }
-
-  var undefinedArgs = [ctorArguments, ctorNamedArguments]
-      .expand((l) => l)
-      .where((pe) => pe.type.isUndefined)
-      .toList();
-  if (undefinedArgs.isNotEmpty) {
-    var description =
-        undefinedArgs.map((fe) => '`${fe.displayName}`').join(', ');
-
-    throw new InvalidGenerationSourceError(
-        'At least one constructor argument has an invalid type: $description.',
-        todo: 'Check names and imports.');
-  }
-
-  // find fields that aren't already set by the constructor and that aren't final
-  var remainingFieldsForFactoryBody = fields.values
-      .where((field) => !field.isFinal)
-      .toSet()
-      .difference(fieldsSetByFactory);
-
-  //
-  // Generate the static factory method
-  //
-  buffer.write('new $className(');
-  buffer.writeAll(
-      ctorArguments.map((paramElement) => deserializeForField(
-          fields[paramElement.name],
-          ctorParam: paramElement)),
-      ', ');
-  if (ctorArguments.isNotEmpty && ctorNamedArguments.isNotEmpty) {
-    buffer.write(', ');
-  }
-  buffer.writeAll(ctorNamedArguments.map((paramElement) {
-    var value =
-        deserializeForField(fields[paramElement.name], ctorParam: paramElement);
-    return '${paramElement.name}: $value';
-  }), ', ');
-
-  buffer.write(')');
-  if (remainingFieldsForFactoryBody.isEmpty) {
-    buffer.writeln(';');
-  } else {
-    for (var field in remainingFieldsForFactoryBody) {
-      buffer.writeln();
-      buffer.write('      ..${field.name} = ');
-      buffer.write(deserializeForField(field));
-      fieldsSetByFactory.add(field);
-    }
-    buffer.writeln(';');
-  }
-  buffer.writeln();
-
-  return fieldsSetByFactory;
 }
